requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.0
scipy>=1.5.0
Pillow>=8.0.0
opencv-python>=4.5.0
albumentations>=1.0.0
scikit-image>=0.18.0
tqdm>=4.50.0
pyyaml>=5.3.0
tensorboardX>=2.2.0
matplotlib>=3.3.0
wandb>=0.12.0
scikit-learn>=0.24.0
einops>=0.3.0
openslide-python>=1.1.1
huggingface-hub>=0.10.0
pytest>=6.0.0
